chore: rename component to use caps filename
