+++ src/pages/messages/[address].tsx
@@ -1,3 +1,3 @@
-import Message from '@components/Messages/message';
+import Message from '@components/Messages/Message';
 
 export default Message;
